requirements.txt
langchain
langchain-anthropic
langchain-ollama
langgraph
langchain-huggingface
langchain-community
langchain-chroma
python-dotenv
pypdf
docx2txt
sentence-transformers
